% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mco_five_meta_analysis.R
\name{mco_five_meta_analysis}
\alias{mco_five_meta_analysis}
\title{MCO one diseases}
\usage{
mco_five_meta_analysis(
  data1 = NULL,
  data2 = NULL,
  data3 = NULL,
  data4 = NULL,
  data5 = NULL,
  NF,
  measurePM_m = NULL,
  q = NULL
)
}
\arguments{
\item{NF}{Number of frontiers Pareto-efficient.}

\item{data}{Gene expression data set.}

\item{measurePM1}{One performance measures (1 = mean, 2 = median, 3 = quantail)}

\item{measurePM2}{Two performance measures (1 = mean, 2 = median, 3 = quantail)}

\item{q1}{if measurePM1 = 3, q1 take the value of the quantile, for examnple q1 = 75.}

\item{q2}{if measurePM2 = 3, q2 take the value of the quantile, for examnple q1 = 95.}
}
\value{
list of genes with maximum expression changes belonging to each Pareto efficient frontier,
a plot of the Pareto efficient frontiers with the maximum expression genes.
}
\description{
This function models a macroarray or RNA-seq gene expression dataset, 
generating a list of genes with maximal changes in expression.
}
\examples{
# Example1: El siguiente ejemplo...
m1 <- mco_one_diseases(data = dataExample, 
                       NF = 5, 
                       measurePM1 = 1, 
                       measurePM2 = 2, 
                       q1 = NULL, 
                       q2 = NULL)
m1$final2

}
\references{
Citar articulo
}
\author{

}
