% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mco_one_diseases_3PM.R
\name{mco_one_diseases_3PM}
\alias{mco_one_diseases_3PM}
\title{mco_one_diseases_3PM Function}
\usage{
mco_one_diseases_3PM(
  data = NULL,
  NF,
  measurePMs31 = NULL,
  measurePMs32 = NULL,
  measurePMs33 = NULL,
  q1 = NULL,
  q2 = NULL,
  q3 = NULL
)
}
\arguments{
\item{data}{Gene expression data set.Gene expression data set.}

\item{NF}{Number of frontiers Pareto-efficient.}

\item{q1}{if measurePM1 = 3, q1 take the value of the quantile, for examnple q1 = 75}

\item{q2}{if measurePM2 = 3, q2 take the value of the quantile, for examnple q1 = 95.}

\item{q3}{if measurePM3 = 3, q3 take the value of the quantile, for examnple q1 = 80.}

\item{measurePM1}{One performance measures (1 = mean, 2 = median, 3 = quantail).}

\item{measurePM2}{Two performance measures (1 = mean, 2 = median, 3 = quantail).}

\item{measurePM3}{Three performance measures (1 = mean, 2 = median, 3 = quantail).}
}
\value{
A  list with 2 elements.
\itemize{
\item \code{GeneFrontier} is a genes list with maximum expression changes belonging to each Pareto efficient frontier.
\item \code{PlotFrontier} is a 3D plot of the Pareto efficient frontiers with the maximum expression genes.

}
}
\description{
This function models a macroarray or RNA-seq gene expression dataset,
generating a list of genes with maximal changes in expression.
}
\examples{
# Example1: El siguiente ejemplo...
m1 <- mco_one_diseases_3PM(data = dataExample,
                       NF = 5,
                       measurePM1 = 1,
                       measurePM2 = 2,
                       measurePM3 = 3,
                       q1 = NULL,
                       q2 = NULL,
                       q3 = NULL)
m1$final2

}
\references{
[1] Camacho-Cáceres, K. I. et al. Multiple criteria optimization joint analyses of microarray experiments in lung cancer: from existing microarray data to new knowledge. Cancer Med. 4, 1884–1900 (2015).
            [2] Narváez-Bandera, I., Suárez-Gómez, D., Isaza, C. E. & Cabrera-Ríos, M. Multiple Criteria Optimization (MCO): A gene selection deterministic tool in RStudio. PLoS One 17, e0262890 (2022).
}
\author{
Deiver Suárez-Gómez [aut],
        Isis Narváez-Bandera [aut],
        Clara E. Isaza1 [aut],
        Mauricio Cabrera-Ríos1 [aut]
}
