% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_MST_one_disease.R
\name{MST_one_disease}
\alias{MST_one_disease}
\title{MST_one_disease Function}
\usage{
MST_one_disease(data_expresio = NULL, data_gene = NULL)
}
\arguments{
\item{data_gene}{list of genes with maximal changes in expression.}

\item{data_expresion}{Gene expression data set.Gene expression data set.}
}
\value{
is a maximum correlation between gene pairs.
}
\description{
This function find the most correlated acyclic path between gene
changes based on the pairwise Pearson correlation coefficient using one omics datasets.
}
\examples{
# Example7: El siguiente ejemplo...
#' m1 <- MST_one_disease(data_expresion = dataExample1,
                       data_gene = geneList)


}
\references{
Isaza, C. et al. Biological signaling pathways and potential mathematical network representations: biological discovery through optimization. Cancer Med. 7, 1875–1895 (2018).
}
\author{
Deiver Suárez-Gómez [aut],
        Isis Narváez-Bandera [aut],
        Clara E. Isaza [aut],
        Mauricio Cabrera-Ríos [aut]
}
