% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_MST_two_disease.R
\name{MST_tow_disease}
\alias{MST_tow_disease}
\title{MST_two_meta_diseases Function}
\usage{
MST_tow_disease(
  data_expresion1 = NULL,
  data_expresion2 = NULL,
  data_gene = NULL
)
}
\arguments{
\item{data_expresion1}{Gene expression data set.Gene expression data set.}

\item{data_expresion2}{Gene expression data set.Gene expression data set.}

\item{data_gene}{list of genes with maximal changes in expression.}
}
\value{
is a maximum correlation between gene pairs.
}
\description{
This function find the most correlated acyclic path between gene
changes based on the pairwise Pearson correlation coefficient using two omics datasets.
}
\examples{
# Example8: El siguiente ejemplo...
#' m1 <- MST_tow_disease(data_expresion1 = dataExample1,
                       data_expresion2 = dataExample2,
                       data_gene = geneList)


}
\references{
Isaza, C. et al. Biological signaling pathways and potential mathematical network representations: biological discovery through optimization. Cancer Med. 7, 1875–1895 (2018).
}
\author{
Deiver Suárez-Gómez [aut],
        Isis Narváez-Bandera [aut],
        Clara E. Isaza1 [aut],
        Mauricio Cabrera-Ríos1 [aut]
}
