% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_OGF_groups.R
\name{OGF_groups}
\alias{OGF_groups}
\title{OGF_groups}
\usage{
OGF_groups(incidence_matrix, gorups, mpg)
}
\arguments{
\item{incidence_matrix}{A binary incidence matrix that relates genes
and biological processes}

\item{gorups}{The maximum number of groups to be formed.}

\item{mpg}{The maximum number of biological processes per group.}
}
\value{
A list of indices defining genes, biological processes, 
and the groups to which they belong.
}
\description{
This function generates groups of genes and biological processes
based on an incidence matrix.
}
