% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mco_one_diseases.R
\name{mco_one_diseases}
\alias{mco_one_diseases}
\title{MCO one diseases}
\usage{
mco_one_diseases(
  data = NULL,
  NF,
  measurePM1 = NULL,
  measurePM2 = NULL,
  q1 = NULL,
  q2 = NULL
)
}
\arguments{
\item{data}{Gene expression data set.}

\item{NF}{Number of frontiers Pareto-efficient.}

\item{measurePM1}{One performance measures (1 = mean, 2 = median, 3 = quantail)}

\item{measurePM2}{Two performance measures (1 = mean, 2 = median, 3 = quantail)}

\item{q1}{if measurePM1 = 3, q1 take the value of the quantile, for examnple q1 = 75.}

\item{q2}{if measurePM2 = 3, q2 take the value of the quantile, for examnple q1 = 95.}
}
\value{
list of genes with maximum expression changes belonging to each Pareto efficient frontier,
a plot of the Pareto efficient frontiers with the maximum expression genes.
}
\description{
This function models a macroarray or RNA-seq gene expression dataset, 
generating a list of genes with maximal changes in expression.
}
\examples{
# Example1: El siguiente ejemplo...
library(OBAMA)
X <- dataExample
m1 <- mco_one_diseases(data = X, 
                       NF = 5, 
                       measurePM1 = 1, 
                       measurePM2 = 2, 
                       q1 = NULL, 
                       q2 = NULL)
m1$final2

}
\references{
[1] Camacho-Cáceres, K. I. et al. Multiple criteria optimization joint analyses of microarray experiments in lung cancer: from existing microarray data to new knowledge. Cancer Med. 4, 1884–1900 (2015).
            [2] Narváez-Bandera, I., Suárez-Gómez, D., Isaza, C. E. & Cabrera-Ríos, M. Multiple Criteria Optimization (MCO): A gene selection deterministic tool in RStudio. PLoS One 17, e0262890 (2022).
}
\author{
Deiver Suárez-Gómez [aut],
        Isis Narváez-Bandera [aut],
        Clara E. Isaza [aut],
        Mauricio Cabrera-Ríos [aut]
}
