% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mco_one_diseases_3PM.R
\name{mco_one_diseases_3PM}
\alias{mco_one_diseases_3PM}
\title{mco_one_diseases_3PM Function}
\usage{
mco_one_diseases_3PM(
  data = NULL,
  NF,
  measurePMs31 = NULL,
  measurePMs32 = NULL,
  measurePMs33 = NULL,
  q1 = NULL,
  q2 = NULL,
  q3 = NULL
)
}
\arguments{
\item{data}{treatments, number of new lines for test.}

\item{NF}{number of checks per augmented block.}

\item{q1}{esto}

\item{q2}{esto...}

\item{measurePM1}{number of blocks (Augmented blocks).}

\item{measurePM2}{serpentine or cartesian. serpentine by default.}

\item{measurePM3}{esto..}
}
\value{
List of parameters with information about of number of blocks, treatments, vector of treatments,
number of sites, plot numbers, locations names and the seed number.

RCBD layout
}
\description{
It generates a augmented randomized complete block design (ARCBD).
}
\examples{
# Example1: El siguiente ejemplo...
m1 <- mco_one_diseases_3PM(data = dataExample, 
                       NF = 5, 
                       measurePM1 = 1, 
                       measurePM2 = 2, 
                       measurePM3 = 3,
                       q1 = NULL, 
                       q2 = NULL,
                       q3 = NULL)
m1$final2

}
\references{
Citar articulo
}
\author{
Deiver Suarez Gomez
}
